cloud microservices/cloud microservices/ocr/requirements.txt
Flask==1.1.2
google-cloud-vision==2.3.0
gunicorn==20.0.4
